An Airplane Plot Demonstrating Migration Events

This function allows for the geographic plotting of clade centers and geographical outliers. Arcs connect clade centers and outliers. This function accepts data from 3 time points.

airplane_plot(df, dfcoor, time1, time2, time3)

Arguments

df

A data frame of samples containing at least columns: 1. clade: A numeric value. df may have no more than 40 clades with no more than 50 individual samples per clade. 2. country: A string value containing the origin of the sample. 3. longitude: A numeric value containing the longitudinal coordinate of the sample's location. 4. latitue: A numeric value containing the latitudinal coordinate of the sample's location. 5. timepoint: An interger equaling 1, 2, or 3. Representing the time point in which the sample was taken.

dfcoor

A data frame clade center locations to be plotted containng at least columns: 1. clade: A numeric value reflecting the clade number cooresponding to the location of the sample. There must be no more or fewer clades than in df. 2. location: A string value containing the location of the center of the clade. 3. longitude: A numeric value containing the longitudinal coordinate of the clade center's location. 3. latitude: A numeric value containing the latitudinal coordinate of the clade center's location.

time1

A string value conataing the name of the first time point. Example: "January", "December 1-20", "2022".

time2

A string value conataing the name of the second time point. Example: "January", "December 1-20", "2022".

time3

A string value conataing the name of the third time point. Example: "January", "December 1-20", "2022".

Value

Ariplane plot with imposed connections between the clade centers and the geographical outliers of the clade. Clade centers are represented on the mercator projection by larger black dots. Geographic outliers are represented by smaller dots. The arcs indicate a connection between the clade center and the geographic outlier. The color of the arcs indicate the time point that the geographical outlier falls under.

Examples

df <- transRate:: sample_df
dfcoor <- transRate:: sample_dfcoor
airplane_plot(df,dfcoor,"January","February","March")
#> Error in requireNamespace(geosphere): object 'geosphere' not found