Simulate Migration Events in the Coalescent

This function allows users to simulate migration events in a coalescent. The migration rate is drawn from a binomial distribution using the migration_rate parameter.

migration_sim(migration_rate, treestr, popnames, popnum)

Arguments

migration_rate

A numerical value containing the desired migration rate.

treestr

A tree string, written in Newick format, containing the coalescent to undergo the migration simulation.

popnames

The names of populations in the tree string as a list.

Value

Matrix containing the transformed populations and ancestral nodes.

Examples

treestr <- "(((H1:0.00402#0.01,C1:0.00402#0.01):0.00304#0.01,G2:0.00707#0.01):0.00929#0.01,C2:0.01635#0.01)#0.01;"
migration_rate <- 0.5
popnames <- c("H","C","G")
popnum <- 3
migration_sim(migration_rate,treestr,popnames,popnum)
#> Error in if (inode > nspecies) {    son1 <- treematrix[inode, 2]    son2 <- treematrix[inode, 3]    treematrix <- noclock2clock(son1, treematrix, nspecies)    treematrix <- noclock2clock(son2, treematrix, nspecies)    leftheight <- treematrix[son1, 4] + node.height(son1, treematrix,         nspecies)    rightheight <- treematrix[son2, 4] + node.height(son2, treematrix,         nspecies)    leftratio <- (leftheight + rightheight)/2/leftheight    rightratio <- (leftheight + rightheight)/2/rightheight    leftsonnodes <- offspring.nodes(son1, treematrix, nspecies)    rightsonnodes <- offspring.nodes(son2, treematrix, nspecies)    treematrix[leftsonnodes, 4] <- treematrix[leftsonnodes, 4] *         leftratio    treematrix[rightsonnodes, 4] <- treematrix[rightsonnodes,         4] * rightratio}: missing value where TRUE/FALSE needed